Move fetchers to folders (#1112)
