+++ resources/fetching/factory/eks_factory.go
@@ -27,7 +27,8 @@ import (
 
 	"github.com/elastic/cloudbeat/resources/fetching"
 	"github.com/elastic/cloudbeat/resources/fetching/condition"
-	"github.com/elastic/cloudbeat/resources/fetching/fetchers"
+	awsfetchers "github.com/elastic/cloudbeat/resources/fetching/fetchers/aws"
+	k8sfetchers "github.com/elastic/cloudbeat/resources/fetching/fetchers/k8s"
 	"github.com/elastic/cloudbeat/resources/providers/awslib"
 	"github.com/elastic/cloudbeat/resources/providers/awslib/ecr"
 	"github.com/elastic/cloudbeat/resources/providers/awslib/elb"
@@ -39,7 +40,7 @@ const (
 )
 
 var (
-	eksRequiredProcesses = fetchers.ProcessesConfigMap{"kubelet": {ConfigFileArguments: []string{"config"}}}
+	eksRequiredProcesses = k8sfetchers.ProcessesConfigMap{"kubelet": {ConfigFileArguments: []string{"config"}}}
 	eksFsPatterns        = []string{
 		"/hostfs/etc/kubernetes/kubelet/kubelet-config.json",
 		"/hostfs/var/lib/kubelet/kubeconfig"}
@@ -52,29 +53,29 @@ func NewCisEksFactory(log *logp.Logger, awsConfig aws.Config, ch chan fetching.R
 	if identity != nil {
 		log.Info("Initialize aws-related fetchers")
 		ecrPrivateProvider := ecr.NewEcrProvider(log, awsConfig, &awslib.MultiRegionClientFactory[ecr.Client]{})
-		privateRepoRegex := fmt.Sprintf(fetchers.PrivateRepoRegexTemplate, identity.Account)
+		privateRepoRegex := fmt.Sprintf(awsfetchers.PrivateRepoRegexTemplate, identity.Account)
 
-		ecrPodDescriber := fetchers.PodDescriber{
+		ecrPodDescriber := awsfetchers.PodDescriber{
 			FilterRegex: regexp.MustCompile(privateRepoRegex),
 			Provider:    ecrPrivateProvider,
 		}
 
-		ecrFetcher := fetchers.NewEcrFetcher(log, ch, k8sClient, ecrPodDescriber)
+		ecrFetcher := awsfetchers.NewEcrFetcher(log, ch, k8sClient, ecrPodDescriber)
 		m[fetching.EcrType] = RegisteredFetcher{Fetcher: ecrFetcher, Condition: []fetching.Condition{condition.NewIsLeader(log, le)}}
 
 		elbProvider := elb.NewElbProvider(awsConfig)
 		loadBalancerRegex := fmt.Sprintf(elbRegexTemplate, awsConfig.Region)
-		elbFetcher := fetchers.NewElbFetcher(log, ch, k8sClient, elbProvider, identity, loadBalancerRegex)
+		elbFetcher := awsfetchers.NewElbFetcher(log, ch, k8sClient, elbProvider, identity, loadBalancerRegex)
 		m[fetching.ElbType] = RegisteredFetcher{Fetcher: elbFetcher, Condition: []fetching.Condition{condition.NewIsLeader(log, le)}}
 	}
 
-	fsFetcher := fetchers.NewFsFetcher(log, ch, eksFsPatterns)
+	fsFetcher := k8sfetchers.NewFsFetcher(log, ch, eksFsPatterns)
 	m[fetching.FileSystemType] = RegisteredFetcher{Fetcher: fsFetcher}
 
-	procFetcher := fetchers.NewProcessFetcher(log, ch, eksRequiredProcesses)
+	procFetcher := k8sfetchers.NewProcessFetcher(log, ch, eksRequiredProcesses)
 	m[fetching.ProcessType] = RegisteredFetcher{Fetcher: procFetcher}
 
-	kubeFetcher := fetchers.NewKubeFetcher(log, ch, k8sClient)
+	kubeFetcher := k8sfetchers.NewKubeFetcher(log, ch, k8sClient)
 	m[fetching.KubeAPIType] = RegisteredFetcher{Fetcher: kubeFetcher, Condition: []fetching.Condition{condition.NewIsLeader(log, le)}}
 	return m
 }
