+++ resources/fetching/fetchers/aws/ecr_fetcher.go
