+++ resources/fetching/fetchers/aws/ecr_fetcher_test.go
