+++ resources/fetching/fetchers/aws/logging_fetcher.go
