+++ resources/fetching/fetchers/aws/logging_fetcher_test.go
