+++ resources/fetching/fetchers/aws/network_fetcher.go
