+++ resources/fetching/fetchers/aws/network_fetcher_test.go
